fix: No more wrong/missing user photos/names
The problem was that if you pass to VkAPI request with multiple userIDs some of which are repeated,
response won't contain all the duplicated user objects, but only unique ones. So when some user wrote
something in two or more dialogs there will be problems in such case, so we need to store sequence of ids.